Add Copernicus Marine WMS Tiles to a leaflet map

Create an interactive map with leaflet::leaflet() and add layers of Copernicus marine WMS data to it.

Usage

addCopernicusWMSTiles(
  map,
  product,
  layer,
  variable,
  options = leaflet::WMSTileOptions(format = "image/png", transparent = TRUE),
  ...
)

Arguments

map

A map widget object created from leaflet::leaflet()

product

An identifier (type character) of the desired Copernicus marine product. Can be obtained with copernicus_products_list.

layer

The name of a desired layer within a product (type character). Can be obtained with copernicus_product_details.

variable

The name of a desired variable in a specific layer of a product (type character). Can be obtained with copernicus_product_details.

options

Passed on to leaflet::addWMSTiles().

...

Passed on to leaflet::addWMSTiles().

Value

Returns an updated map

Note

WMS functions don't work on systems that don't support GDAL utils

See also

Author

Pepijn de Vries

Examples

if (FALSE) {
if (interactive()) {
  leaflet::leaflet() |>
    leaflet::setView(lng = 3, lat = 54, zoom = 4) |>
    leaflet::addProviderTiles("Esri.WorldImagery") |>
    addCopernicusWMSTiles(
      product  = "GLOBAL_ANALYSISFORECAST_PHY_001_024",
      layer    = "cmems_mod_glo_phy-thetao_anfc_0.083deg_P1D-m",
      variable = "thetao")
}
}