Contact Copernicus Marine login page

Contact Copernicus Marine login page and check if login is successful.

Usage

cms_login(
  username = getOption("CopernicusMarine_uid", ""),
  password = getOption("CopernicusMarine_pwd", "")
)

Arguments

username

Your Copernicus marine user name. Can be provided as options(CopernicusMarine_uid = "my_user_name"), or as argument here.

password

Your Copernicus marine password. Can be provided as options(CopernicusMarine_pwd = "my_password"), or as argument here.

Value

Returns a named list with your account details if successful, returns NULL otherwise.

Details

This function will return your account details if successful.

Author

Pepijn de Vries

Examples

if (FALSE) { # \dontrun{
## This will return FALSE if you have not set your account details with 'options'.
## If you have specified your account details and there are no other problems,
## it will return TRUE.
cms_login()
} # }