Obtain available services for a specific Copernicus marine product

Obtain an overview of services provided by Copernicus for a specific marine product.

Usage

cms_product_services(product)

Arguments

product

An identifier (type character) of the desired Copernicus marine product. Can be obtained with cms_products_list.

Value

Returns a tibble with a list of available services for a Copernicus marine product.

See also

Author

Pepijn de Vries

Examples

cms_product_services("GLOBAL_ANALYSISFORECAST_PHY_001_024")
#> # A tibble: 23 × 4
#>    layer                                             ext   `WWW:STAC` `OGC:WMTS`
#>    <chr>                                             <chr> <list>     <list>    
#>  1 cmems_mod_glo_phy_anfc_0.083deg_static            coor… <chr [1]>  <NULL>    
#>  2 cmems_mod_glo_phy_anfc_0.083deg_static            NA    <NULL>     <chr [3]> 
#>  3 cmems_mod_glo_phy_anfc_0.083deg_static            bathy <chr [1]>  <NULL>    
#>  4 cmems_mod_glo_phy_anfc_0.083deg_static            mdt   <chr [1]>  <NULL>    
#>  5 cmems_mod_glo_phy_anfc_merged-uv_PT1H-i           NA    <chr [1]>  <chr [1]> 
#>  6 cmems_mod_glo_phy_anfc_0.083deg-climatology-unce… NA    <chr [1]>  <chr [1]> 
#>  7 cmems_mod_glo_phy_anfc_0.083deg_P1D-m             NA    <chr [1]>  <chr [1]> 
#>  8 cmems_mod_glo_phy_anfc_0.083deg_P1M-m             NA    <chr [1]>  <chr [1]> 
#>  9 cmems_mod_glo_phy_anfc_0.083deg_PT1H-m            NA    <chr [1]>  <chr [1]> 
#> 10 cmems_mod_glo_phy-so_anfc_0.083deg_P1D-m          NA    <chr [1]>  <chr [1]> 
#> # ℹ 13 more rows