Deprecated

Full marine data sets can be downloaded using the SpatioTemporal Asset Catalogs (STAC). The API for interacting with these files is no longer operational. Please use cms_download_native() instead.

Usage

cms_download_stac(
  file_tibble,
  destination,
  show_progress = TRUE,
  overwrite = FALSE
)

cms_list_stac_files(product, layer)

cms_stac_properties(product, layer)

Arguments

file_tibble

Ignored

destination

Ignored

show_progress

Ignored

overwrite

A logical value. When FALSE (default), files at the destination won't be overwritten when the exist. Instead an error will be thrown if this is the case. When set to TRUE, existing files will be overwritten.

product

An identifier (type character) of the desired Copernicus marine product. Can be obtained with cms_products_list.

layer

The name of a desired layer within a product (type character). Can be obtained with cms_product_details.

Value

returns NULL, as the API is no longer available

Author

Pepijn de Vries